requirements.txt
streamlit>=1.30.0
openai>=1.12.0
openpyxl>=3.1.2
pandas>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
